requirements.txt
openai
anthropic
google-generativeai
requests
colorama